requirements.txt
mysql-connector-python==8.1.0
neo4j==5.12.0